# docs/api/exports/frontend/openapi.yaml
openapi: 3.0.1
info:
  title: KTTKPM_Nhom15_BE API
  description: Professional OpenAPI 3 documentation for the KTTKPM_Nhom15_BE backend.
  contact:
    name: Backend Team
    email: backend-team@example.local
  license:
    name: Internal Use
    url: https://example.local/internal-use
  version: v1
...
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/v1/admin/variants/{id}:
    put:
      tags:
      - Admin Catalog
      operationId: updateVariant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VariantSummaryResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/v1/admin/products/{id}:
    get:
      tags:
      - Admin Catalog
      operationId: getProduct_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
      - Admin Catalog
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/v1/products:
    get:
      tags:
      - Catalog
      operationId: getProducts
      parameters:
      - name: typeId
        in: query
        required: false
        schema:
          type: string
      - name: minPrice
        in: query
        required: false
        schema:
          type: number
      - name: maxPrice
        in: query
        required: false
        schema:
          type: number
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProductSummaryDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      deprecated: true
    post:
      tags:
      - Catalog
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompositeProductRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreatedResourceResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      deprecated: true
...
  /api/v1/options:
    get:
      tags:
      - Catalog
      operationId: getOptions
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OptionResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      deprecated: true
    post:
      tags:
      - Catalog
      operationId: createOption
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      deprecated: true
...
  /api/v1/admin/products:
    get:
      tags:
      - Admin Catalog
      operationId: getProducts_1
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: typeId
        in: query
        required: false
        schema:
          type: string
      - name: active
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProductListItemResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
      - Admin Catalog
      operationId: createProduct_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreatedResourceResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/v1/admin/options:
    get:
      tags:
      - Admin Catalog
      operationId: getOptions_1
      parameters:
      - name: active
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageOptionListItemResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
      - Admin Catalog
      operationId: createOption_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionWriteRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreatedResourceResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/v1/products/{productId}/variants/{variantId}:
    patch:
      tags:
      - Catalog
      operationId: patchVariant
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      - name: variantId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchVariantRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      deprecated: true
...
  /api/v1/products/{id}:
    get:
      tags:
      - Catalog
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      deprecated: true
...
components:
  schemas:
...
    OptionRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        values:
          type: array
          items:
            type: string
    OptionResponse:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        values:
          type: array
          items:
            $ref: "#/components/schemas/OptionValueResponse"
...
    VariantUpdateRequest:
      required:
      - isActive
      - price
      - sku
      - stockQuantity
      type: object
      properties:
        sku:
          type: string
        price:
          minimum: 0
          type: number
        stockQuantity:
          minimum: 0
          type: integer
          format: int32
        isActive:
          type: boolean
        options:
          type: array
          items:
            $ref: "#/components/schemas/VariantOptionAssignmentRequest"
...
    VariantSummaryResponse:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        normalizedSku:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
          format: int32
        active:
          type: boolean
        options:
          type: array
          items:
            $ref: "#/components/schemas/OptionAssignmentResponse"
...
    ProductUpdateRequest:
      required:
      - isActive
      - isCustomizable
      - name
      - typeId
      type: object
      properties:
        typeId:
          type: string
        name:
          type: string
        slug:
          type: string
        descriptionMd:
          type: string
        shortDescription:
          type: string
        brandId:
          type: string
        isCustomizable:
          type: boolean
        isActive:
          type: boolean
        ingredientIds:
          type: array
          items:
            type: string
        skinTypeIds:
          type: array
          items:
            type: string
        concernIds:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: string
    ProductDetailResponse:
      type: object
      properties:
        id:
          type: string
        typeId:
          type: string
        name:
          type: string
        slug:
          type: string
        descriptionMd:
          type: string
        shortDescription:
          type: string
        brandId:
          type: string
        brandName:
          type: string
        customizable:
          type: boolean
        active:
          type: boolean
        minPrice:
          type: number
        maxPrice:
          type: number
        totalStock:
          type: integer
          format: int64
        activeVariantCount:
          type: integer
          format: int64
        reviewCount:
          type: integer
          format: int64
        averageRating:
          type: number
        soldCount:
          type: integer
          format: int64
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/SemanticReferenceResponse"
        skinTypes:
          type: array
          items:
            $ref: "#/components/schemas/SemanticReferenceResponse"
        concerns:
          type: array
          items:
            $ref: "#/components/schemas/SemanticReferenceResponse"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/SemanticReferenceResponse"
        searchStatus:
          $ref: "#/components/schemas/SearchStatusResponse"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantSummaryResponse"
...
    OptionWriteRequest:
      required:
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
        values:
          type: array
          items:
            $ref: "#/components/schemas/OptionValueWriteRequest"
...
    CompositeProductRequestDTO:
      required:
      - name
      - typeId
      - variants
      type: object
      properties:
        typeId:
          type: string
        name:
          type: string
        descriptionMd:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantRequestDTO"
        customizable:
          type: boolean
...
    CreatedResourceResponse:
      type: object
      properties:
        id:
          type: string
...
    ProductCreateRequest:
      required:
      - isActive
      - isCustomizable
      - name
      - typeId
      - variants
      type: object
      properties:
        typeId:
          type: string
        name:
          type: string
        slug:
          type: string
        descriptionMd:
          type: string
        shortDescription:
          type: string
        brandId:
          type: string
        isCustomizable:
          type: boolean
        isActive:
          type: boolean
        ingredientIds:
          type: array
          items:
            type: string
        skinTypeIds:
          type: array
          items:
            type: string
        concernIds:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantCreateRequest"
...
    PageProductSummaryDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummaryDTO"
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
...
    PageProductListItemResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductListItemResponse"
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
...
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorDetail"